complete5 done, changed rank average to float

+++ src/names/Complete4.java
@@ -1,8 +1,8 @@
 package names;
 
 public class Complete4 {
-    public static int run(String name, String gender, int[] years) throws Exception {
-        int ret = Main.findAverageRank(years,gender,name);
+    public static float run(String name, String gender, int[] years) throws Exception {
+        float ret = Main.findAverageRank(years,gender,name);
         return ret;
     }
 }
