+++ src/names/Main.java
@@ -12,8 +12,8 @@ public class Main {
 
     }
 
-    public static int findAverageRank(int[] years, String gender, String name) throws Exception {
-        int totalRank = 0;
+    public static float findAverageRank(int[] years, String gender, String name) throws Exception {
+        float totalRank = 0;
 
         if(years[1]<years[0]){
             throw new Exception("Invalid Year Range");
@@ -24,7 +24,7 @@ public class Main {
             int rank = n.findRank(name);
             totalRank += rank;
         }
-        int avgRank = totalRank/(years[1]-years[0]+1);
+        float avgRank = totalRank/(years[1]-years[0]+1);
 
         return avgRank;
     }
